+++ sentry/src/main/java/io/sentry/protocol/OperatingSystem.java
@@ -35,7 +35,7 @@ public final class OperatingSystem implements IUnknownPropertiesConsumer, Clonea
   private Boolean rooted;
 
   @SuppressWarnings("unused")
-  private Map<String, Object> unknown;
+  private Map<String, @NotNull Object> unknown;
 
   public String getName() {
     return name;
@@ -92,7 +92,7 @@ public final class OperatingSystem implements IUnknownPropertiesConsumer, Clonea
 
   @ApiStatus.Internal
   @Override
-  public void acceptUnknownProperties(Map<String, Object> unknown) {
+  public void acceptUnknownProperties(Map<String, @NotNull Object> unknown) {
     this.unknown = new ConcurrentHashMap<>(unknown);
   }