+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Feat: Add option to ignore exceptions by type (#1352)
 * Feat: Sentry closes Android NDK and ShutdownHook integrations (#1358)
 * Enhancement: Allow inheritance of SentryHandler class in sentry-jul package(#1367)
+* Fix: Accept only non null value maps (#1368)
 
 # 4.4.0-alpha.1